+++ pkg/upload/android-proguard.go
@@ -28,7 +28,6 @@ func ProcessAndroidProguard(options options.CLI, logger log.Logger) error {
 	proguardOptions := options.Upload.AndroidProguard
 
 	var mappingFile string
-	var appManifestPathExpected string
 	var err error
 
 	for _, path := range proguardOptions.Path {
@@ -57,10 +56,13 @@ func ProcessAndroidProguard(options options.CLI, logger log.Logger) error {
 
 			// Attempt to locate AndroidManifest.xml for the variant if not set
 			if proguardOptions.AppManifest == "" {
-				appManifestPathExpected = filepath.Join(path, "app", "build", "intermediates", "merged_manifests", proguardOptions.Variant, "AndroidManifest.xml")
-				if utils.FileExists(appManifestPathExpected) {
-					proguardOptions.AppManifest = appManifestPathExpected
-					logger.Info(fmt.Sprintf("Found app manifest at: %s", proguardOptions.AppManifest))
+				appBuildPath := filepath.Join(path, "app", "build")
+				proguardOptions.AppManifest, err = android.FindAndroidManifest(appBuildPath, proguardOptions.Variant)
+				if err != nil {
+					return err
+				}
+				if proguardOptions.AppManifest == "" {
+					logger.Info("Unable to locate AndroidManifest.xml for variant; proceeding without manifest path")
 				}
 			}
 
@@ -70,15 +72,14 @@ func ProcessAndroidProguard(options options.CLI, logger log.Logger) error {
 
 			// Try to find manifest and variant if missing
 			if proguardOptions.AppManifest == "" && proguardOptions.Variant == "" {
-				mergedManifestPath := filepath.Join(path, "..", "..", "..", "..", "intermediates", "merged_manifests")
-				if filepath.Base(mergedManifestPath) == "merged_manifests" {
-					proguardOptions.Variant, err = android.GetVariantDirectory(mergedManifestPath)
-					if err == nil {
-						appManifestPathExpected = filepath.Join(mergedManifestPath, proguardOptions.Variant, "AndroidManifest.xml")
-						if utils.FileExists(appManifestPathExpected) {
-							proguardOptions.AppManifest = appManifestPathExpected
-							logger.Info(fmt.Sprintf("Found app manifest at: %s", proguardOptions.AppManifest))
-						}
+				appBuildPath := filepath.Join(path, "..", "..", "..", "..")
+				if filepath.Base(appBuildPath) == "build" {
+					proguardOptions.AppManifest, err = android.FindAndroidManifest(appBuildPath, proguardOptions.Variant)
+					if err != nil {
+						return err
+					}
+					if proguardOptions.AppManifest == "" {
+						logger.Info("Unable to locate AndroidManifest.xml for variant; proceeding without manifest path")
 					}
 				}
 			}