+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## unreleased
+
+### Changed
+
+- Updated the logic for the `upload js` command to ignore the top level `node_modules` directory when processing source maps. [#251](https://github.com/bugsnag/bugsnag-cli/pull/251)
+
 ## [3.5.1] - 2025-11-13
 
 ### Changed