+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Updated the logic for the `upload js` command to ignore the top level `node_modules` directory when processing source maps. [#251](https://github.com/bugsnag/bugsnag-cli/pull/251)
+- Add proxy support to the NPM package. [#252](https://github.com/bugsnag/bugsnag-cli/pull/252)
 
 ## [3.5.1] - 2025-11-13