+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## unreleased
+
+### Changed
+
+- Add a fallback path for AndroidManifest.xml when processing Android related uploads (Proguard, NDK and React Native) [#249](https://github.com/bugsnag/bugsnag-cli/pull/249)
+
 ## [3.5.1] - 2025-11-13
 
 ### Changed